requirements/addition_requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
importlib-metadata==1.7.0
lxml==4.5.2
numpy==1.19.1
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.10
djangorestframework-jwt==1.11.0
drf_yasg==1.17.1